modified
imported modules have been given abbreviated names to make them more readable.
Unnecessary printouts have been commented out.
It is made so that dictionaries are entered in the list, if there are more dictionaries. And then in the "reading_TimeTable()" method, the "Type" is checked, and based on that, it is read and printed in the console.